[MOV] mail: move component hooks up
Task-2759571
