+++ addons/mail/static/src/component_hooks/use_component_to_model.js
