+++ addons/mail/static/src/component_hooks/use_drag_visible_dropzone.js
