+++ addons/mail/static/src/component_hooks/use_models.js
