+++ addons/mail/static/src/component_hooks/use_ref_to_model.js
