+++ addons/mail/static/src/component_hooks/use_refs.js
