+++ addons/mail/static/src/component_hooks/use_rendered_values.js
