+++ addons/mail/static/src/component_hooks/use_update.js
