+++ addons/mail/static/src/components/attachment_viewer/attachment_viewer.js
@@ -1,8 +1,8 @@
 /** @odoo-module **/
 
+import { useComponentToModel } from '@mail/component_hooks/use_component_to_model';
+import { useRefs } from '@mail/component_hooks/use_refs';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
-import { useComponentToModel } from '@mail/component_hooks/use_component_to_model/use_component_to_model';
-import { useRefs } from '@mail/component_hooks/use_refs/use_refs';
 import { hidePDFJSButtons } from '@web/legacy/js/libs/pdfjs';
 
 const { Component, onMounted, onPatched, onWillUnmount, useRef } = owl;
