+++ addons/mail/static/src/components/channel_invitation_form/channel_invitation_form.js
@@ -1,9 +1,9 @@
 /** @odoo-module **/
 
-import { useComponentToModel } from '@mail/component_hooks/use_component_to_model/use_component_to_model';
+import { useComponentToModel } from '@mail/component_hooks/use_component_to_model';
+import { useRefToModel } from '@mail/component_hooks/use_ref_to_model';
+import { useUpdateToModel } from '@mail/component_hooks/use_update_to_model';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
-import { useRefToModel } from '@mail/component_hooks/use_ref_to_model/use_ref_to_model';
-import { useUpdateToModel } from '@mail/component_hooks/use_update_to_model/use_update_to_model';
 
 const { Component } = owl;
