+++ addons/mail/static/src/components/chatter/chatter.js
@@ -1,8 +1,8 @@
 /** @odoo-module **/
 
+import { useComponentToModel } from '@mail/component_hooks/use_component_to_model';
+import { useUpdate } from '@mail/component_hooks/use_update';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
-import { useComponentToModel } from '@mail/component_hooks/use_component_to_model/use_component_to_model';
-import { useUpdate } from '@mail/component_hooks/use_update/use_update';
 import { LegacyComponent } from '@web/legacy/legacy_component';
 
 const { useRef } = owl;
