+++ addons/mail/static/src/components/message_action_list/message_action_list.js
@@ -1,7 +1,7 @@
 /** @odoo-module */
 
+import { useRefToModel } from '@mail/component_hooks/use_ref_to_model';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
-import { useRefToModel } from '@mail/component_hooks/use_ref_to_model/use_ref_to_model';
 
 const { Component } = owl;
