+++ addons/mail/static/src/components/popover_view/popover_view.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { useComponentToModel } from '@mail/component_hooks/use_component_to_model/use_component_to_model';
+import { useComponentToModel } from '@mail/component_hooks/use_component_to_model';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
 
 import { usePosition } from '@web/core/position_hook';
