+++ addons/mail/static/src/components/rtc_call_viewer/rtc_call_viewer.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { useUpdate } from '@mail/component_hooks/use_update/use_update';
+import { useUpdate } from '@mail/component_hooks/use_update';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
 
 import Dialog from 'web.OwlDialog';
