+++ addons/mail/static/src/components/rtc_layout_menu/rtc_layout_menu.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
+import { useComponentToModel } from '@mail/component_hooks/use_component_to_model';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
-import { useComponentToModel } from '@mail/component_hooks/use_component_to_model/use_component_to_model';
 import { LegacyComponent } from "@web/legacy/legacy_component";
 
 export class RtcLayoutMenu extends LegacyComponent {
