+++ addons/mail/static/src/components/rtc_video/rtc_video.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
+import { useUpdate } from '@mail/component_hooks/use_update';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
-import { useUpdate } from '@mail/component_hooks/use_update/use_update';
 
 const { Component } = owl;
