+++ addons/mail/static/src/components/thread_view_topbar/thread_view_topbar.js
@@ -1,8 +1,8 @@
 /** @odoo-module **/
 
+import { useRefToModel } from '@mail/component_hooks/use_ref_to_model';
+import { useUpdateToModel } from '@mail/component_hooks/use_update_to_model';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
-import { useRefToModel } from '@mail/component_hooks/use_ref_to_model/use_ref_to_model';
-import { useUpdateToModel } from '@mail/component_hooks/use_update_to_model/use_update_to_model';
 
 const { Component } = owl;
