+++ addons/mail/static/src/utils/messaging_component.js
@@ -1,6 +1,6 @@
 /** @odoo-module */
 
-import { useModels } from "@mail/component_hooks/use_models/use_models";
+import { useModels } from '@mail/component_hooks/use_models';
 
 const { useRef } = owl;
